requirements.txt

playwright
opencv-python
pillow
numpy
python-dotenv   # opcional, para .env

